Fix UB, cap quadratic diff memory, and clean all strict-build warnings
- web.cpp/auth.hpp/server_stateless.hpp/fs_helpers.hpp: fix undefined
  behavior from passing signed char to std::tolower and range-for over a
  char view as unsigned char. The is_blocked_host() SSRF guard and the
  search-result URL dedup were affected on non-ASCII bytes.
- diff.cpp: guard the O(n*m) LCS matrix with a 16M-cell cap; above it,
  fall back to a correct full-replace edit script instead of allocating
  ~100MB+ (and hanging) on multi-MB inputs. Rewrote compute_edits with
  size_t indices.
- Drive -Wconversion/-Wsign-conversion/-Wmissing-field-initializers to
  zero across lib, tests, and examples so the tree builds clean under
  -Werror: explicit casts in fs/search size formatters, repomap iterator
  offset, sha256 loop indices; default member initializers on Request,
  SubprocessOptions, Implementation, ServerCapabilities.

Builds clean under strict warnings + -Werror; all 17 tests pass. Diff
guard verified under ASan/UBSan and a 60MB vmem cap.

+++ src/tools/repomap.cpp
@@ -613,7 +613,7 @@ mcp::cap::Result run_repo_map(const Json& args) {
         // source order (rank chose WHICH to show; line chooses the order).
         constexpr std::size_t kCap = 24;
         std::vector<const Def*> pick(ds.begin(),
-            ds.begin() + std::min<std::size_t>(kCap, ds.size()));
+            ds.begin() + static_cast<std::ptrdiff_t>(std::min<std::size_t>(kCap, ds.size())));
         std::sort(pick.begin(), pick.end(),
                   [](const Def* x, const Def* y) { return x->line < y->line; });
         for (const auto* d : pick) {
